paineluniversal/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
websockets==12.0
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
aioredis==2.0.1

# QR Code generation
qrcode[pil]==7.4.2
Pillow==10.1.0

# Validação de documentos brasileiros
# python-cpf-cnpj==1.0.0  # Package not available, commented out

# Excel e relatórios
openpyxl==3.1.2
xlsxwriter==3.1.9
reportlab==4.0.7

# PDF generation
weasyprint==60.2

# Email
emails==0.6.0
jinja2==3.1.2

# HTTP requests
httpx==0.25.2
aiohttp==3.9.1

# Configuração e environment
pydantic-settings==2.1.0

# Utilitários de data e tempo
python-dateutil==2.8.2
pytz==2023.3

# Validação e formatação
validators==0.22.0
phonenumbers==8.13.25

# Logging estruturado
structlog==23.2.0

# Monitoramento e métricas
prometheus-client==0.19.0

# Criptografia avançada
cryptography==41.0.7

# Processamento de imagens
opencv-python-headless==4.8.1.78

# Análise de dados
pandas==2.1.4
numpy==1.25.2

# Geração de gráficos
matplotlib==3.8.2
plotly==5.17.0

# Machine Learning (para IA avançada)
scikit-learn==1.3.2

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
